+++ src/events/event-types.js
@@ -8,8 +8,6 @@ export const Events = {
   EQUATION_TOGGLED: 'equation:toggled',
   EQUATIONS_RESET: 'equations:reset',
   THEME_CHANGED: 'theme:changed',
-  BREAKPOINTS_CHANGED: 'breakpoints:changed',
-  CAST_TIMES_CHANGED: 'cast-times:changed',
   TAB_CHANGED: 'tab:changed',
 
   // Group management events